requirements.txt
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.5
schedule==1.2.0
tinydb==4.8.0 